Add a special platform name 'unix'
Darwin has derived from Unix and Linux has developed based on (or to mimic) Unix.
So most of dotfiles can be used in both Linux/Darwin and I could not be botherd to keep
mappings_linux.json and mappings_darwin.json equal.

This PR add a special platform named 'unix'. In Unix-like platform, this special platform
name will be used prior to the specific one ('linux' or 'darwin') and the specific one will be used
posterior (they override the config for 'unix').